+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 * Upgrade `frequenz-client-reporting` to minimum `v0.16.0`.
 
 ## New Features
+* Notification service: Introduced `NotificationSendError` for structured retry failure handling and updated the logging in `send_test_email()` utility function.
 
 ## Bug Fixes